# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: ProxyLoom P0 API
  version: 1.0.0
  description: >-
    P0 contract for a single configured workspace and trusted administrators.
    Five administrator authentication operations are implemented. Other operations
    remain contract-only; declarations do not assert kernel capability or verification.
    Management uses the session Cookie; internal operations exist exclusively on
    a separate mTLS listener; subscription retrieval uses its mandatory path token
    independently of either credential. Scope and resource IDs are server assigned.
    JSON decoding rejects unknown fields, duplicate keys, trailing documents,
    invalid Unicode, and undocumented nulls. Errors never echo input values.
    All references in this document are local; schema validation must not fetch URLs.
servers:
  - url: /
    description: Management and public subscription listener; internal paths use their own listener.
security:
  - SessionCookie: []
...
components:
  securitySchemes:
    SessionCookie:
      type: "apiKey"
      in: "cookie"
      name: "proxyloom_session"
      description: "Opaque server-side administrator session; Secure, HttpOnly, SameSite=Lax. State-changing management operations additionally enforce Origin and X-CSRF-Token. Setup/login require Origin and JSON but no session CSRF token. Secure may be omitted only for explicit development HTTP loopback PUBLIC_URL. Never accepted as subscription or Runner authorization."
    RunnerMTLS:
      type: "mutualTLS"
      description: "Registered client certificate verified only on the dedicated internal listener. Certificate identity and server registration determine eligible builds, task types and slots. No management Cookie or subscription token is accepted."
...
  parameters:
...
    ID:
      name: "id"
      in: "path"
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
    IfMatch:
      name: "If-Match"
      in: "header"
      required: true
      description: "Exactly one strong tag \"r<N>\" for the resource being changed. Missing => 428; malformed, weak, wildcard, multi-tag or overflow => 400; valid stale revision => 412. Business/generation/lease/idempotency conflicts => 409. Creates and authentication operations have no resource precondition."
      schema:
        type: "string"
        pattern: "^\"r(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)\"$"
...
    Limit:
      name: "limit"
      in: "query"
      schema:
        type: "integer"
        minimum: 1
        maximum: 200
        default: 50
      description: "Bounded page size. Stable ascending (created_at, id) order unless a collection documents its immutable equivalent."
    Cursor:
      name: "cursor"
      in: "query"
      schema:
        type: "string"
        minLength: 1
        maxLength: 2048
      description: "Opaque authenticated continuation tied to scope, collection, order and normalized filter digest. A changed filter or invalid cursor is 400; authorization is checked separately on every request."
    Enabled:
      name: "enabled"
      in: "query"
      schema:
        type: "boolean"
    Tag:
      name: "tag"
      in: "query"
      schema: { $ref: "#/components/schemas/Tag" }
    Protocol:
      name: "protocol"
      in: "query"
      schema: { $ref: "#/components/schemas/Protocol" }
    NodeSearch:
      name: "q"
      in: "query"
      schema:
        type: "string"
        maxLength: 256
      description: "Case-insensitive node name search; secrets are never searched or echoed."
...
  responses:
    BadRequest:
      description: "Malformed JSON, duplicate/unknown field, unsafe shape, invalid path/query/header, invalid Unicode or cursor. No submitted key or value is echoed."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: "Required management session or registered internal identity is missing, invalid or expired; surfaces are not interchangeable."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: "Authenticated actor lacks permission, recent authentication, a valid current-actor confirmation or required same-origin protection."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: "Resource does not exist in the authorized scope, or existence must not be disclosed."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: "Business state, generation, idempotency or lease conflict, including COMPILE_OBSOLETE, IDEMPOTENCY_CONFLICT and LEASE_LOST. A stale resource revision is 412 instead."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PreconditionFailed:
      description: "A syntactically valid If-Match names an outdated resource revision; no mutation was applied."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    TooLarge:
      description: "Input, decompressed data, item count or event size exceeds the configured bound."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnprocessableEntity:
      description: "Shape is understood but merged domain semantics are invalid, including required secret clear, unsupported capability, invalid references, excluded dependencies and DNS cycles."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PreconditionRequired:
      description: "The versioned resource operation requires If-Match but it was omitted."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Unavailable:
      description: "Primary database, execution capability or safe publication is unavailable. Fail closed; never serve cached authorization or substitute a direct/empty configuration."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NodeResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/NodeReadResponse" }
...
    NodeBatchResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/NodeBatchResponse" }
    NodeListResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/NodeListResponse" }
...
  schemas:
...
    NodeCreateRequest:
      type: "object"
      additionalProperties: false
      required: ["name", "node"]
      properties:
        name: { $ref: "#/components/schemas/Name" }
        tags: { $ref: "#/components/schemas/Tags" }
        enabled:
          type: "boolean"
        node: { $ref: "#/components/schemas/NodeCreate" }
    NodePatchRequest:
      type: "object"
      additionalProperties: false
      required: []
      properties:
        name: { $ref: "#/components/schemas/Name" }
        tags: { $ref: "#/components/schemas/Tags" }
        enabled:
          type: "boolean"
        node: { $ref: "#/components/schemas/NodePatch" }
      minProperties: 1
...
paths:
...
  /api/v1/nodes:
    get:
      operationId: listNodes
      x-implementation-status: implemented
      summary: List nodes using management-only redacted DTOs
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Enabled'}, {$ref: '#/components/parameters/Tag'}, {$ref: '#/components/parameters/Protocol'}, {$ref: '#/components/parameters/NodeSearch'}]
      responses:
        '200': {$ref: '#/components/responses/NodeListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/Unavailable'}
    post:
      operationId: createNode
      x-implementation-status: implemented
      summary: Create a typed node at revision 1
      description: Generates ID, scope, revision and epoch on the server. No If-Match is required for creation.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NodeCreateRequest'}
      responses:
        '201': {$ref: '#/components/responses/NodeResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/nodes/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getNode
      x-implementation-status: implemented
      summary: Read a node without returning auth or REALITY secrets
      responses:
        '200': {$ref: '#/components/responses/NodeResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
    patch:
      operationId: updateNode
      x-implementation-status: implemented
      summary: Apply a typed node patch and create a new immutable revision
      description: >-
        An absent secret preserves its previous value, null clears it, and a
        string replaces it. The merged node is validated; clearing a required
        password, UUID, username or REALITY field returns 422. Empty or masked
        credentials are invalid. Authentication changes and disabling advance
        the security epoch; a caller cannot label a security change as ordinary.
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NodePatchRequest'}
      responses:
        '200': {$ref: '#/components/responses/NodeResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '503': {$ref: '#/components/responses/Unavailable'}
    delete:
      operationId: deleteNode
      x-implementation-status: implemented
      summary: Soft-delete a node and immediately revoke dependent publication access
      description: Preserves immutable history and references; advances revision and security epoch.
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      responses:
        '200': {$ref: '#/components/responses/MutationResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/nodes/batch:
    post:
      operationId: updateNodesBatch
      x-implementation-status: implemented
      summary: Apply a bounded metadata operation with individual revision preconditions
      description: Each node has its own precondition and result. A missing item precondition is 428 and a stale one is 412 in that item's result; the batch has no synthetic shared ETag.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NodeBatchRequest'}
      responses:
        '200': {$ref: '#/components/responses/NodeBatchResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/nodes/{id}/revisions:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: listNodeRevisions
      x-implementation-status: implemented
      summary: List immutable redacted revisions with their original metadata
      description: Historical name, tags, enabled and epoch are taken from that revision, never from the current head. History cannot authorize publication of old secrets.
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Cursor'}]
      responses:
        '200': {$ref: '#/components/responses/NodeListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/nodes/{id}/references:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: listNodeReferences
      x-implementation-status: implemented
      summary: List current active and historical reverse references
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - name: reference_state
          in: query
          schema: {type: string, enum: [active, historical, all], default: active}
      responses:
        '200': {$ref: '#/components/responses/ReferenceListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/nodes/{id}/reveal:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      operationId: revealNodeSecrets
      x-implementation-status: implemented
      summary: Reveal typed secrets once after recent authentication with an audit record
      description: Requires the matched node revision and server-verified recent authentication. This dedicated response must never be cached, logged or stored for idempotent replay.
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      responses:
        '200': {$ref: '#/components/responses/NodeRevealResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/nodes/{id}/clone:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      operationId: cloneNode
      x-implementation-status: implemented
      summary: Clone the matched node revision with a new stable ID
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NodeCloneRequest'}
      responses:
        '201': {$ref: '#/components/responses/NodeResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '503': {$ref: '#/components/responses/Unavailable'}